requirements.txt
aiogram==3.4.1
google-generativeai==0.5.4
aiosqlite==0.20.0
python-dotenv==1.0.1